fix: move express server to server.ts
The motivation behind this is to let index.ts handle any preparation
work (database setup) before instantiating the HTTP servers.
